Rename pyflatsurf_conversion to pyflatsurf.conversion
